+++ pro23/src/com/spring/ex03/MemberDAO.java
@@ -1,54 +1,37 @@
-package com.spring.ex03;
-
-import java.io.Reader;
-import java.util.List;
-
-import org.apache.ibatis.io.Resources;
-import org.apache.ibatis.session.SqlSession;
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.apache.ibatis.session.SqlSessionFactoryBuilder;
-
-import com.spring.ex01.MemberVO;
-
-public class MemberDAO {
-	public static SqlSessionFactory sqlMapper = null;
-
-	private static SqlSessionFactory getInstance() {
-		if (sqlMapper == null) {
-			try {
-				String resource = "mybatis/SqlMapConfig.xml";
-				Reader reader = Resources.getResourceAsReader(resource);
-				sqlMapper = new SqlSessionFactoryBuilder().build(reader);
-				reader.close();
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-		return sqlMapper;
-	}
-	public List<MemberVO> selectAllMemberList() {
-		sqlMapper = getInstance();
-		SqlSession session = sqlMapper.openSession();
-		List<MemberVO> memlist = null;
-		memlist = session.selectList("mapper.member.selectAllMemberList");
-		return memlist;
-	}
-
-	public MemberVO selectMemberById(String id){
-	      sqlMapper=getInstance();
-	SqlSession session=sqlMapper.openSession();
-	      MemberVO memberVO=session.selectOne("mapper.member.selectMemberById",id);
-	      return memberVO;		
-	   }
-
-	public List<MemberVO> selectMemberByPwd(int pwd) {
-	sqlMapper = getInstance();
-	SqlSession session = sqlMapper.openSession();
-	List<MemberVO> membersList = null;
-	membersList= session.selectList("mapper.member.selectMemberByPwd", pwd);
-	return membersList;
-	}
-
-
-
-}
+/*
+ * package com.spring.ex03;
+ * 
+ * import java.io.Reader; import java.util.List;
+ * 
+ * import org.apache.ibatis.io.Resources; import
+ * org.apache.ibatis.session.SqlSession; import
+ * org.apache.ibatis.session.SqlSessionFactory; import
+ * org.apache.ibatis.session.SqlSessionFactoryBuilder;
+ * 
+ * import com.spring.ex01.MemberVO;
+ * 
+ * public class MemberDAO { public static SqlSessionFactory sqlMapper = null;
+ * 
+ * private static SqlSessionFactory getInstance() { if (sqlMapper == null) { try
+ * { String resource = "mybatis/SqlMapConfig.xml"; Reader reader =
+ * Resources.getResourceAsReader(resource); sqlMapper = new
+ * SqlSessionFactoryBuilder().build(reader); reader.close(); } catch (Exception
+ * e) { e.printStackTrace(); } } return sqlMapper; } public List<MemberVO>
+ * selectAllMemberList() { sqlMapper = getInstance(); SqlSession session =
+ * sqlMapper.openSession(); List<MemberVO> memlist = null; memlist =
+ * session.selectList("mapper.member.selectAllMemberList"); return memlist; }
+ * 
+ * public MemberVO selectMemberById(String id){ sqlMapper=getInstance();
+ * SqlSession session=sqlMapper.openSession(); MemberVO
+ * memberVO=session.selectOne("mapper.member.selectMemberById",id); return
+ * memberVO; }
+ * 
+ * public List<MemberVO> selectMemberByPwd(int pwd) { sqlMapper = getInstance();
+ * SqlSession session = sqlMapper.openSession(); List<MemberVO> membersList =
+ * null; membersList= session.selectList("mapper.member.selectMemberByPwd",
+ * pwd); return membersList; }
+ * 
+ * 
+ * 
+ * }
+ */
