+++ pro23/src/com/spring/sdo0911/MemberDAO.java
@@ -8,6 +8,8 @@ import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.apache.ibatis.session.SqlSessionFactoryBuilder;
 
+import com.spring.ex01.MemberVO;
+
 public class MemberDAO {
 	public static SqlSessionFactory sqlMapper = null;
 
@@ -23,9 +25,7 @@ public class MemberDAO {
 			}
 		}
 		return sqlMapper;
-
 	}
-//원본
 	public List<MemberVO> selectAllMemberList() {
 		sqlMapper = getInstance();
 		SqlSession session = sqlMapper.openSession();
@@ -33,21 +33,37 @@ public class MemberDAO {
 		memlist = session.selectList("mapper.member.selectAllMemberList");
 		return memlist;
 	}
-	//과제
-		public List<MemberVO> selectAllMemberList2() {
-			sqlMapper = getInstance();
-			SqlSession session = sqlMapper.openSession();
-			List<MemberVO> memlist = null;
-			memlist = session.selectList("mapper.member.selectAllMemberList2");
-			return memlist;
-		}
+
+	public MemberVO selectMemberById(String id){
+	      sqlMapper=getInstance();
+	      SqlSession session=sqlMapper.openSession();
+	      MemberVO memberVO=session.selectOne("mapper.member.selectMemberById",id);
+	      return memberVO;		
+	   }
+
+	public List<MemberVO> selectMemberByPwd(int pwd) {
+		sqlMapper = getInstance();
+		SqlSession session = sqlMapper.openSession();
+		List<MemberVO> membersList = null;
+		membersList= session.selectList("mapper.member.selectMemberByPwd", pwd);
+		return membersList;
+	}
 	
-//	 public List<HashMap<String, String>> selectAllMemberList() { 
-//		sqlMapper = getInstance(); 
-//     	SqlSession session = sqlMapper.openSession();
-//		List<HashMap<String, String>> memlist = null; 
-//   	memlist = session.selectList("mapper.member.selectAllMemberList"); 
-//		return memlist; 
-//	 }
+	public List<MemberVO> selectMemberByName(String id){
+		sqlMapper = getInstance();
+		SqlSession session = sqlMapper.openSession();
+		List<MemberVO> memlist = null;
+		memlist = session.selectList("mapper.member.selectAllMemberByName", id);
+		return memlist;
+	   }
 	
+	public MemberVO selectMemberByEmail(String id){
+	      sqlMapper=getInstance();
+	      SqlSession session=sqlMapper.openSession();
+	      MemberVO memberVO=session.selectOne("mapper.member.selectMemberByEmail",id);
+	      return memberVO;		
+	   }
+
+
+
 }
